Type planet owners

+++ src/yt_game.h
@@ -85,7 +85,7 @@ struct yt_planet {
 	float production[3];
 	float stock[3];
 	float missiles;
-	float owner;
+	int owner;
 	float ground_forces;
 	size_t name_length;
 	float last_minute;
@@ -464,7 +464,7 @@ bool yt_planet_landing_unrest_row(float reduced, float original,
 void yt_planet_assault_player_overlay(struct yt_player *player,
     float commitment);
 void yt_planet_assault_victory_overlay(struct yt_planet *planet,
-    float owner, float attackers);
+	int owner, float attackers);
 void yt_planet_assault_failure_overlay(struct yt_planet *planet,
     float defenders);
 void yt_planet_assault_round(bool attacker_damage, float amount,
@@ -724,16 +724,16 @@ bool yt_projectile_victim_mines_overlay(struct yt_player *player,
 bool yt_projectile_sector_mines_overlay(struct yt_sector *sector,
     float carried_mines);
 bool yt_projectile_planet_ground_overlay(struct yt_planet *planet,
-    float ground, float owner);
+    float ground, int owner);
 bool yt_projectile_planet_productivity_overlay(struct yt_planet *planet,
     const float production[3], const float stock[3]);
 bool yt_projectile_planet_destroy_overlay(struct yt_planet *planet);
 bool yt_projectile_sector_unlink_overlay(struct yt_sector *sector);
 struct yt_projectile_ground_result {
 	float ground;
-	float owner;
+	int owner;
 };
-bool yt_projectile_planet_ground_damage(float ground, float owner,
+bool yt_projectile_planet_ground_damage(float ground, int owner,
     float *remaining, struct yt_random *random,
     struct yt_projectile_ground_result *result, struct yt_error *error);
 struct yt_projectile_productivity_result {
